[NG] Refactor api-select into api folder
