+++ packages/ng/demo/app/api/api-select/api-select.module.ts
@@ -3,6 +3,7 @@ import { CommonModule } from '@angular/common';
 import { HttpClientModule} from '@angular/common/http';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { LuSelectModule } from '../../../../src/app/select';
+import { LuSelectApiModule } from '../../../../src/app/api/select';
 import { SharedModule } from '../../shared/index';
 import { BasicApiSelectComponent } from './basic/basic';
 import { DemoApiSelectFeeder } from './basic/api-select-feeder';
@@ -18,6 +19,7 @@ import { DemoApiSelectComponent } from './api-select.component';
 		HttpClientModule,
 		ReactiveFormsModule,
 		LuSelectModule,
+		LuSelectApiModule,
 		SharedModule,
 	],
 	declarations: [
