+++ packages/ng/demo/app/api/api-select/basic/basic.html
@@ -1,5 +1,5 @@
 <section class="contentSection">
-	<h3>Component With api select</h3>
+	<h3>Component With api select load by module</h3>
 	<div class="textfield">
 		<lu-api-select [(ngModel)]="itemSelect"
 			[selectApiFeeder]="apiSelectFeeder"
@@ -12,7 +12,7 @@
 	<code class="code mod-block">{{ itemSelect | json }}</code>
 </section>
 <section class="contentSection">
-	<h3>Component With api select</h3>
+	<h3>Component With api select load by component</h3>
 	<div class="textfield">
 		<lu-api-select [(ngModel)]="itemSelect"
 			[selectApiFeeder]="feeder"
