--- packages/ng/demo/app/api/picker/api-picker.component.html
@@ -1,15 +1,0 @@
-<demo-api-docs directive="LuApiPickerDirective"></demo-api-docs>
-<demo-api-docs directive="LuApiPickerComponent"></demo-api-docs>
-<demo-redirect></demo-redirect>
-<demo-example-box [snippets]="snippets" demo="basic">
-	<h2>Basic</h2>
-	<demo-api-basic></demo-api-basic>
-</demo-example-box>
-<!-- <demo-example-box [snippets]="snippets" demo="nopicker">
-	<h2>Without picker</h2>
-	<p>
-		you can only use the directive luApiPicker on a tag input
-		and it will try to match what you are typing to a unique item available via this api
-	</p>
-	<demo-api-nopicker></demo-api-nopicker>
-</demo-example-box> -->
