--- packages/ng/demo/app/api/picker/api-picker.component.ts
@@ -1,24 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-import { HttpClient } from '@angular/common/http';
-declare var require: any;
-
-@Component({
-	selector: 'demo-api-picker',
-	templateUrl: './api-picker.component.html',
-	styles: []
-})
-export class DemoApiPickerComponent implements OnInit {
-	snippets = {
-		basic: {
-			code: require('!!prismjs-loader?lang=typescript!./basic/basic'),
-			markup: require('!!prismjs-loader?lang=markup!./basic/basic.html')
-		},
-		nopicker: {
-			code: require('!!prismjs-loader?lang=typescript!./nopicker/nopicker'),
-			markup: require('!!prismjs-loader?lang=markup!./nopicker/nopicker.html')
-		},
-	};
-	constructor(private http: HttpClient) { }
-	ngOnInit() {
-	}
-}
