--- packages/ng/demo/app/api/picker/nopicker/nopicker.html
@@ -1,12 +1,0 @@
-<!-- <div class="textfield mod-material">
-	<input [(ngModel)]="api" class="textfield-input" (ngModelChange)="apiChanged()" name="api" luEmpty>
-	<label class="textfield-label" for="api">api</label>
-</div>
-<div class="textfield mod-material">
-	<input luApiPicker [(ngModel)]="item" class="textfield-input" name="item" luEmpty [api]="api">
-	<label class="textfield-label" for="item">item</label>
-	<div class="textfield-messages">
-		<span class="textfield-messages-helper">type at least 2 letters</span>
-	</div>
-</div>
-<code class="code mod-block">{{ item | json }}</code> -->
