--- packages/ng/demo/app/api/picker/nopicker/nopicker.ts
@@ -1,12 +1,0 @@
-import { Component } from '@angular/core';
-@Component({
-	selector: 'demo-api-nopicker',
-	templateUrl: './nopicker.html',
-})
-export class NopickerComponent {
-	api = '/api/v3/axissections';
-	item = undefined;
-	apiChanged() {
-		this.item = undefined;
-	}
-}
