+++ packages/ng/demo/app/select/select.router.ts
@@ -1,6 +1,5 @@
 import { Routes, Route } from '@angular/router';
 import { DemoSelectComponent } from './select.component';
-import { DemoApiSelectComponent } from './api-select/api-select.component';
 import { DemoClearerSelectComponent } from './clearer-select/clearer-select.component';
 import { DemoModSelectComponent } from './mod-select/mod-select.component';
 import { DemoFeederSelectComponent } from './feeder-select/feeder-select.component';
@@ -8,7 +7,6 @@ import { DemoSearcherSelectComponent } from './searcher-select/searcher-select.c
 
 export const selectRoutes: Routes = [
 	{ path: 'select', label: 'Select', children: [
-		{ path: 'api', label: 'Api', component: DemoApiSelectComponent } as Route,
 		{ path: 'clear', label: 'Clear', component: DemoClearerSelectComponent } as Route,
 		{ path: 'mod', label: 'Mods', component: DemoModSelectComponent } as Route,
 		{ path: 'feeder', label: 'Feeder', component: DemoFeederSelectComponent } as Route,
