+++ packages/ng/src/app/api/api.module.ts
@@ -1,20 +1,20 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { FormsModule } from '@angular/forms';
-import { LuApiPickerModule } from './picker';
 import { HttpClientModule } from '@angular/common/http';
+import {LuSelectApiModule} from './select';
 
 @NgModule({
 	imports: [
 		CommonModule,
 		FormsModule,
 		HttpClientModule,
-		LuApiPickerModule
+		LuSelectApiModule,
 	],
 	declarations: [
 	],
 	exports: [
-		LuApiPickerModule,
+		LuSelectApiModule,
 	],
 })
 export class LuApiModule { }
