+++ packages/ng/src/app/api/index.ts
@@ -1,3 +1,3 @@
 export { LuApiModule } from './api.module';
 export { ICoerce, IApiItem } from './api.model';
-export { LuApiPickerModule, LuApiPickerComponent, LuApiPickerDirective, customApiPickerTemplate } from './picker';
+export * from './select';
