--- packages/ng/src/app/api/picker/api-picker.component.html
@@ -1,17 +1,0 @@
-<ng-template>
-	<div class="lu-popover-panel" role="dialog" [class.lu-popover-overlap]="overlapTrigger" [ngClass]="_classList" [ngStyle]="popoverPanelStyles"
-	 (keydown)="_handleKeydown($event)" (click)="onClick()" (mouseover)="onMouseOver()" (mouseleave)="onMouseLeave()" (mousedown)="onMouseDown($event)"
-	 [@transformPopover]="'enter'">
-		<div class="lu-popover-content" [ngStyle]="popoverContentStyles" cdkTrapFocus="focusTrapEnabled">
-			<ul class="api-picker-options">
-				<li
-				*ngFor="let option of (options$ | async); let i = index"
-				(click)="selectOption(option)"
-				[ngClass]="{'is-focus': (i === (highlightIndex$ | async))}"
-				class="api-picker-option">
-					<span>{{ option.name }}</span>
-				</li>
-			</ul>
-		</div>
-	</div>
-</ng-template>
