--- packages/ng/src/app/api/picker/api-picker.component.scss
@@ -1,25 +1,0 @@
-@import "../../popover/popover.component";
-
-.lu-popover-content {
-	padding: 0;
-}
-
-.api-picker-options {
-	margin: 0;
-	padding: 0;
-	list-style: none;
-}
-
-.api-picker-option {
-	cursor: pointer;
-	padding: .5rem 1rem;
-	transition: background 50ms ease;
-
-	&:hover {
-		background: _theme("commons.background.dark");
-	}
-
-	&.is-focus {
-		background: _theme("commons.background.dark");
-	}
-}
