--- packages/ng/src/app/api/picker/index.ts
@@ -1,4 +1,0 @@
-export { LuApiPickerModule } from './api-picker.module';
-export { LuApiPickerDirective } from './api-picker.directive';
-export { LuApiPickerComponent } from './api-picker.component';
-export { customApiPickerTemplate, standardApiPickerTemplate } from './api-picker.template';
