+++ packages/ng/src/app/api/select/feeder/api-feeder.module.ts
@@ -1,0 +1,10 @@
+import { NgModule } from '@angular/core';
+
+import { RDDApiFeederComponent } from './api-feeder.component';
+
+@NgModule({
+	imports: [],
+	declarations: [RDDApiFeederComponent],
+	exports: [RDDApiFeederComponent],
+})
+export class LuApiFeederModule {}
