+++ packages/ng/src/app/api/select/picker/select-api-picker.component.scss
