+++ packages/ng/src/app/api/select/picker/select-api-picker.module.ts
@@ -4,7 +4,8 @@ import { FormsModule } from '@angular/forms';
 import { CommonModule } from '@angular/common';
 import { HttpClientModule } from '@angular/common/http';
 import { LuSelectApiPicker } from './select-api-picker.component';
-import { LuSelectOptionModule } from '../../option';
+import { LuSelectOptionModule } from '../../../select';
+import { LuApiFeederModule } from '../feeder';
 
 @NgModule({
 	imports: [
@@ -12,6 +13,7 @@ import { LuSelectOptionModule } from '../../option';
 		FormsModule,
 		BrowserModule,
 		HttpClientModule,
+		LuApiFeederModule,
 		LuSelectOptionModule,
 	],
 	declarations: [
