+++ packages/ng/src/app/formly/formly.config.ts
@@ -5,7 +5,7 @@ import { LuFormlyFieldDate } from './types/date';
 import { LuFormlyFieldTextarea } from './types/textarea';
 import { LuFormlyFieldAutocomplete } from './types/autocomplete';
 import { LuFormlyFieldSelect } from './types/select';
-import { LuFormlyFieldApi } from './types/api';
+// import { LuFormlyFieldApi } from './types/api';
 // wrappers
 import { LuFormlyWrapperHelper, TemplateHelper } from './wrappers/helper';
 import { LuFormlyWrapperTitle, TemplateTitle } from './wrappers/title';
@@ -25,7 +25,7 @@ export const LU_FORMLY_COMPONENTS = [
 	LuFormlyFieldTextarea,
 	LuFormlyFieldAutocomplete,
 	LuFormlyFieldSelect,
-	LuFormlyFieldApi,
+	// LuFormlyFieldApi,
 
 	LuFormlyWrapperHelper,
 	LuFormlyWrapperLabel,
@@ -64,11 +64,11 @@ export const LU_FORMLY_CONFIG = {
 			component: LuFormlyFieldSelect,
 			wrappers: ['textfield-layout'],
 		},
-		{
+		/* {
 			name: 'api',
 			component: LuFormlyFieldApi,
 			wrappers: ['textfield-layout'],
-		},
+		},*/
 	],
 	wrappers: [
 		{ name: 'title', component: LuFormlyWrapperTitle },
