+++ packages/ng/src/app/formly/formly.module.ts
@@ -5,7 +5,6 @@ import { FormlyModule } from '@ngx-formly/core';
 import { MatAutocompleteModule, MatInputModule, MatOptionModule, MatSelectModule, MatDatepickerModule } from '@angular/material';
 import { LU_FORMLY_COMPONENTS, LU_FORMLY_CONFIG } from './formly.config';
 import { LuEmptyModule } from '../empty/empty.module';
-import { LuApiPickerModule } from '../api/picker/api-picker.module';
 
 @NgModule({
 	declarations: [
@@ -21,7 +20,6 @@ import { LuApiPickerModule } from '../api/picker/api-picker.module';
 		MatSelectModule,
 		MatDatepickerModule,
 
-		LuApiPickerModule,
 
 		LuEmptyModule,
