--- packages/ng/src/app/select/api/feeder/api-feeder.module.ts
@@ -1,10 +1,0 @@
-import { NgModule } from '@angular/core';
-
-import { RDDApiFeederComponent } from './api-feeder.component';
-
-@NgModule({
-	imports: [],
-	declarations: [RDDApiFeederComponent],
-	exports: [RDDApiFeederComponent],
-})
-export class LuApiFeederModule {}
