+++ packages/ng/src/app/user/select/user-select-api-feeder.ts
@@ -1,4 +1,4 @@
-import { ISelectApiFeeder } from '../../select';
+import { ISelectApiFeeder } from '../../api';
 import { Observable } from 'rxjs/Observable';
 import { HttpClient } from '@angular/common/http';
 import { Injectable } from '@angular/core';
